Implement Weather API Client with Types and Tests
This commit introduces a simple Weather API client, leveraging the
frequenz-client-base-python framework.

The client is designed to handle requests consisting of location and
feature object lists. It efficiently communicates with the server
using the frequenz-client-base-python, receiving a grpc stream from
the weather service.

Additionally, the client features a receiver that filters the stream
as required and organizes the data into a 3D numpy array. This setup
ensures efficient handling and structuring of weather data.

Comprehensive tests have been added to validate the functionality of
the types defined for this implementation. Tests for the client still
need to be implemented.

- Utilizes frequenz-client-base-python for server communication
- Receives and processes grpc stream from weather service
- Filters stream data based on demand
- Structures data into a 3D numpy array
- Includes tests for validation of new types

Signed-off-by: noah-kreutzer-frequenz <noah.kreutzer@frequenz.com>
Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>