Enhance CI Docker Image for frequenz-client-base-python Integration
This commit updates the Dockerfile in `.github/containers/test-installation` to
prepare for the integration of the `frequenz-client-base-python` repository in
upcoming features. The modifications expand the CI test Docker image to support
this specific repository's requirements.

File modified:
- `.github/containers/test-installation/Dockerfile`

The enhancement of the Docker image is a critical step in aligning our CI
testing environment with the needs of the `frequenz-client-base-python`
repository. This repository will play a significant role in future developments,
and its integration requires a Docker image equipped with additional tools and
dependencies for Git repository management. By updating the Dockerfile, we
ensure that our CI tests remain effective and are fully prepared for the
complexities and functionalities introduced by integrating the
`frequenz-client-base-python` repo. This proactive update is essential for
maintaining the efficiency and accuracy of our CI processes in light of these
new developments.

Signed-off-by: noah-kreutzer-frequenz <noah.kreutzer@frequenz.com>