Add refined version for the to_ndarray_vfl() function #77
According to issue #77 a refined version of to_ndarray_vfl() to_ndarray_vfl_refined() was added to the repo.
to_ndarray_vfl_refined() in comparision to the base function takes the validity_times to filter for
in form of a list of datetime objects instead of a list of timedelta relative to the request.

Signed-off-by: noah-kreutzer-frequenz <noah.kreutzer@frequenz.com>